Moved code
I created a new post object because the ServerRequestHandler class was
getting overwhelmingly large. I’m pretty confident that this is an OK
move because most of the world modifications (save delete) happens in
the post requests. It might even make sense to handle delete in the new
object so that all edits to the world occur there. I’ll probably make a
get object too.